// src/api/generated/models/apimedal.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface APIMedal
 */
export interface APIMedal {
    /**
     * 
     * @type {string}
     * @memberof APIMedal
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof APIMedal
     */
    'image'?: string;
    /**
     * 
     * @type {number}
     * @memberof APIMedal
     */
    'count': number;
}